% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossScreening.R
\docType{package}
\name{CrossScreening-package}
\alias{CrossScreening-package}
\title{Cross-screening for observational studies}
\description{
Cross-screening is a new method that uses both random halves of the sample to screen and test many hypotheses. It generally improves statistical power in observational studies when many hypotheses are tested simultaneously.
}

