% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fallback.test}
\alias{fallback.test}
\title{Fallback procedure for multiple testing}
\usage{
fallback.test(p, alpha = 0.05, spread = 1)
}
\arguments{
\item{p}{a vector of p-values}

\item{alpha}{significance level}

\item{spread}{the way to spread \code{alpha}, either a vector of the same length as \code{p} or a single number to indicate equal spread in the first \code{spread} hypotheses.}
}
\value{
the rejected hypotheses
}
\description{
Fallback procedure for multiple testing
}
\author{
Qingyuan Zhao
}
\references{
Brian L. Wiens. A fixed sequence Bonferroni procedure for testing multiple endpoints. Pharmaceutical Statistics, 2(3), 211---215, 2003.
}

