% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gamma.to.kappa}
\alias{gamma.to.kappa}
\title{Transform a sensitivity parameter from gamma scale to kappa scale}
\usage{
\method{gamma}{to.kappa}(gamma)
}
\description{
Transform a sensitivity parameter from gamma scale to kappa scale
}
\keyword{internal}

