% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{kappa.to.gamma}
\alias{kappa.to.gamma}
\title{Transform a sensitivity parameter from kappa scale to gamma scale}
\usage{
\method{kappa}{to.gamma}(kappa)
}
\description{
Transform a sensitivity parameter from kappa scale to gamma scale
}
\keyword{internal}

