% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{lead}
\alias{lead}
\title{Lead in children}
\format{A data.frame.}
\usage{
data(lead)
}
\description{
Morton et al. (1982) compared the levelsof lead in the blo od of 33 children wh ose fathers worked in afactory w here lead was used in the manufacture of batteriesto 33 lead levels in ma tched control ch ild ren of the sam e agefrom the same neighborhood. The variables are as follows:
}
\details{
\describe{
\item{control}{lead levels (ug/dl)}
\item{level}{father's potential exposure}
\item{hyg}{hygine of father employed in the lead industry}
}
}
\references{
Morton, D. E., Saah, A. J., Silberg, S. L., Owens, W. L., ROBERTS, M. A., & Saah, M. D. (1982). Lead absorption in children of employees in a lead-related industry. American Journal of Epidemiology, 115(4), 549-555.
}
\keyword{datasets}

