% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{nhanes.fish.match}
\alias{nhanes.fish.match}
\title{Pair matching result}
\format{A data.frame.}
\usage{
data(nhanes.fish.match)
}
\description{
Each row is a matched pair, the first/second entry is the id of low/high fish intake in the \code{nhanes.fish} data frame.
}
\keyword{datasets}

