% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{nhanes.log2diff}
\alias{nhanes.log2diff}
\title{Obtains treatment-minus-control differences in the \code{nhanes.fish} dataset}
\usage{
nhanes.log2diff()
}
\value{
a 234 * 46 matrix of the log2 differences
}
\description{
Obtains treatment-minus-control differences in the \code{nhanes.fish} dataset
}

