% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{recycle.test}
\alias{recycle.test}
\title{Recycling procedure for multiple testing}
\usage{
recycle.test(p, alpha = 0.05)
}
\arguments{
\item{p}{a vector of p-values}

\item{alpha}{significance level}
}
\value{
rejected hypotheses
}
\description{
Recycling procedure for multiple testing
}
\details{
WARNING: only supports recycle the first two tests.
}
\author{
Qingyuan Zhao
}

