% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{table5}
\alias{table5}
\title{Generate Table 5 of the paper}
\usage{
table5(no.sims = 1)
}
\arguments{
\item{no.sims}{number of simulations}
}
\description{
Generate Table 5 of the paper
}
\details{
Be careful, this starts a parallel program that may use all your CPU!
}
\keyword{internal}

